refactor(types): widen VShardId and WAL record_type from u16 to u32
Move the canonical `VShardId` definition into `nodedb-types` so it is
available to all crates without re-exporting from the core binary.
The inner representation changes from `u16` to `u32` to accommodate
future growth beyond the current 1024-shard ceiling and to make the
type consistent with `vshard_id` fields in the WAL record header,
which also widens from u16 to u32.

WAL format version bumps from 3 to 4:
- `record_type` field: u16 → u32 (repr change propagated to RecordType)
- `vshard_id` field: u16 → u32
- 16 reserved bytes added to the fixed header (covered by CRC32C)
- HEADER_SIZE grows from 30 to 50 bytes

All call sites across nodedb-cluster, nodedb-bridge, nodedb-wal,
and the core binary are updated to use `u32` throughout.

+++ nodedb-bridge/src/envelope.rs
@@ -14,7 +14,7 @@ pub struct Request {
     pub tenant_id: u32,
 
     /// Virtual shard this request targets.
-    pub vshard_id: u16,
+    pub vshard_id: u32,
 
     /// Opaque execution plan digest (interpreted by the Data Plane).
     pub plan: Plan,
